requirements.txt
PyYAML==5.4.1
PyYAML==6.0.1
Requests==2.31.0
